+++ CHANGELOG.md
@@ -18,7 +18,7 @@ Procedure when bumping the version number:
 - Config file
 - `data_dir` config option
 - `ephemeral` config option
-- `offline` config option
+- `offline` config option and `--offline` cli flag
 - `euph.rooms.<name>.autojoin` config option
 - `euph.rooms.<name>.username` config option
 - `euph.rooms.<name>.force_username` config option